aceptacion valores double el campo dinamico en la tabla de proyecciones

+++ isc.time.report.be/isc.time.report.be.domain/Models/Request/Projections/UpdateProjectionHoursProjectRequest.cs
@@ -12,7 +12,7 @@ namespace isc.time.report.be.domain.Models.Request.Projections
         public string ResourceName { get; set; }
         public decimal HourlyCost { get; set; }
         public int ResourceQuantity { get; set; }
-        public List<int> TimeDistribution { get; set; }
+        public List<double> TimeDistribution { get; set; }
         public decimal TotalTime { get; set; }
         public decimal ResourceCost { get; set; }
         public decimal ParticipationPercentage { get; set; }
